devtools/client/webconsole/new-console-output/test/components/test_message-container.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for MessageContainer component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for MessageContainer component</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");

  const { MessageContainer } = require("devtools/client/webconsole/new-console-output/components/message-container");
  const { ConsoleApiCall } = require("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");
  const { EvaluationResult } = require("devtools/client/webconsole/new-console-output/components/message-types/evaluation-result");
  const { PageError } = require("devtools/client/webconsole/new-console-output/components/message-types/page-error");

  yield testFullRender();
  yield testGetMessageComponent();

  SimpleTest.finish();

  /**
   * Test that passing in a message correctly wires up all the children.
   *
   * The different combinations of children are tested in separate per-component
   * tests. This test just ensures that this component pipes data to its children.
   */
  function testFullRender() {
    const testValue = testCommands.get("console.log()");
    const packet = yield getPacket(testValue.command, testValue.commandType);
    const message = prepareMessage(packet);
    const props = {
      message
    };
    const rendered = renderComponent(MessageContainer, props);

    ok(rendered.textContent.includes(testValue.expectedText),
      "MessageContainer pipes data to its children as expected");
  }

  /**
   * Test that getMessageComponent() returns correct component for each message type.
   */
  function testGetMessageComponent() {
    const testValues = [
      {
        commandObj: testCommands.get("console.log()"),
        expectedComponent: ConsoleApiCall
      },
      {
        commandObj: testCommands.get("new Date()"),
        expectedComponent: EvaluationResult
      },
      {
        commandObj: testCommands.get("pageError"),
        expectedComponent: PageError
      }
    ];

    for (let testValue of testValues) {
      const { commandObj, expectedComponent } = testValue;
      const packet = yield getPacket(commandObj.command, commandObj.commandType);
      const message = prepareMessage(packet);
      const rendered = shallowRenderComponent(MessageContainer, {message});
      is(rendered.type, expectedComponent,
        `MessageContainer nests ${expectedComponent} based on command: ${commandObj.command}`);
    }
  }
});
</script>
</body>
</html>
